Move timed-out stage termination beside its siblings
`_terminate_timed_out_stages` was placed in `_pipeline_internals` next to its
caller, but it is process-lifecycle work: it belongs with `_terminate_process`,
which it delegates to, and `_terminate_pipeline_remaining_stages`, the fail-fast
teardown it mirrors. Moving it there also takes `_pipeline_internals` back
under the 400-line module ceiling once the stacked observability branch lands
its own additions on top — this branch alone was under, the combination was not.

The signature now takes a plain sequence of processes rather than the pipeline
spawn result, so the lifecycle module does not need to know that shape.
