+++ cuprum/_process_lifecycle.py
@@ -250,6 +250,32 @@ def _stages_to_terminate(
     ]
 
 
+async def _terminate_timed_out_stages(
+    processes: cabc.Sequence[asyncio.subprocess.Process],
+    cancel_grace: float,
+) -> None:
+    """Terminate every still-running stage after a pipeline deadline expires.
+
+    ``_wait_for_pipeline`` normally performs this teardown itself when the
+    deadline cancels it. A non-positive deadline gives ``asyncio.wait_for`` a
+    zero timeout, and it then cancels that coroutine before it ever runs, so
+    nothing terminates the stages and they are left orphaned for the rest of
+    their natural life. Terminating here covers that route.
+
+    Running on both routes is safe and deliberate: ``_terminate_process``
+    returns immediately for a stage that has already exited. Doing it before
+    the output gather also lets the stage pipes reach EOF, so a captured
+    pipeline cannot block waiting on a producer that is still running.
+
+    Failures are absorbed — this runs while a timeout is already propagating,
+    and must not replace the ``TimeoutExpired`` the caller is waiting for.
+    """
+    await asyncio.gather(
+        *(_terminate_process(process, cancel_grace) for process in processes),
+        return_exceptions=True,
+    )
+
+
 async def _terminate_pipeline_remaining_stages(
     processes: list[asyncio.subprocess.Process],
     wait_tasks: list[asyncio.Task[int]],
